modified point pillars for use with argoverse dataset

+++ torchplus/train/checkpoint.py
@@ -104,7 +104,8 @@ def save(model_dir,
                 min_step = min([get_step(name) for name in all_ckpts])
                 ckpt_to_delete = "{}-{}.tckpt".format(model_name, min_step)
                 all_ckpts.remove(ckpt_to_delete)
-            os.remove(str(Path(model_dir) / ckpt_to_delete))
+            if os.path.exists(str(Path(model_dir) / ckpt_to_delete)):
+                os.remove(str(Path(model_dir) / ckpt_to_delete))
         all_ckpts_filename = _ordered_unique([Path(f).name for f in all_ckpts])
         ckpt_info_dict['all_ckpts'][model_name] = all_ckpts_filename
         with open(ckpt_info_path, 'w') as f:
@@ -142,6 +143,7 @@ def try_restore_latest_checkpoints(model_dir, models):
     name_to_model = _get_name_to_model_map(models)
     for name, model in name_to_model.items():
         latest_ckpt = latest_checkpoint(model_dir, name)
+        print(latest_ckpt)
         if latest_ckpt is not None:
             restore(latest_ckpt, model)
